flask_frontend/requirements.txt
tqdm==4.41.1
Flask==1.1.2
APScheduler==3.6.3
feedparser==5.2.1
requests==2.24.0
opml==0.5
python_dateutil==2.8.1